requirements.txt
streamlit
google-cloud-speech
google-cloud-storage
google-cloud-language
sounddevice
scipy
numpy
